requirements.txt
fastapi                 0.115.2
prisma                  0.15.0
pydantic                2.9.2
pydantic_core           2.23.4
pypng                   0.20220715.0
PyQRCode                1.2.1
python-dotenv           1.0.1
pytz                    2024.2
razorpay                1.4.2
torch                   2.5.0
uvicorn                 0.32.0
weasyprint              62.3
disutils
asyncpg






pip install speechrecognition
pip install pyaudio
pip install pyttsx3